repaired unit tests (broke in prior change) [no ci]
